Serialization & Cookies
Default layout structure is in its own JSON file which gets loaded with an XMR call.
Layout and its blockers can be serialized. After edit mode, layouts will be saved to cookies. Page attempts to load layout from cookies and falls back on server JSON load.
Settings added to reset the layout to default or log a savable version of the layout.